wip refactor package printer-legacy
