+++ packages/printer-legacy/src/input.js
@@ -1,0 +1,10 @@
+const { printObjectMetadata, printCodeType } = require("./object");
+
+const printInputMetadata = printObjectMetadata;
+
+const printCodeInput = (type) => printCodeType(type, "input");
+
+module.exports = {
+  printInputMetadata,
+  printCodeInput,
+};
