+++ packages/printer-legacy/src/interface.js
@@ -1,0 +1,10 @@
+const { printObjectMetadata, printCodeType } = require("./object");
+
+const printInterfaceMetadata = printObjectMetadata;
+
+const printCodeInterface = (type) => printCodeType(type, "interface");
+
+module.exports = {
+  printInterfaceMetadata,
+  printCodeInterface,
+};
