+++ packages/printer-legacy/src/scalar.js
@@ -1,0 +1,12 @@
+printScalarMetadata = (type) => {
+  return printSpecification(type);
+};
+
+const printCodeScalar = (type) => {
+  return `scalar ${getTypeName(type)}`;
+};
+
+module.exports = {
+  printScalarMetadata,
+  printCodeScalar,
+};
